refactor(cmd/artifact): move artifact subcommands from cmd to internal/artifact
