+++ internal/artifact/search/artifact_search.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cmd
+package search
 
 import (
 	"context"
@@ -20,6 +20,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/falcosecurity/falcoctl/internal/config"
 	"github.com/falcosecurity/falcoctl/internal/utils"
 	"github.com/falcosecurity/falcoctl/pkg/index"
 	"github.com/falcosecurity/falcoctl/pkg/options"
@@ -70,12 +71,12 @@ func NewArtifactSearchCmd(ctx context.Context, opt *options.CommonOptions) *cobr
 }
 
 func (o *artifactSearchOptions) RunArtifactSearch(ctx context.Context, args []string) error {
-	indexConfig, err := index.NewConfig(indexesFile)
+	indexConfig, err := index.NewConfig(config.IndexesFile)
 	if err != nil {
 		return err
 	}
 
-	mergedIndexes, err := utils.Indexes(indexConfig, falcoctlPath)
+	mergedIndexes, err := utils.Indexes(indexConfig, config.FalcoctlPath)
 	if err != nil {
 		return err
 	}
