+++ internal/config/config.go
@@ -1,0 +1,36 @@
+// Copyright 2022 The Falco Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package config
+
+import (
+	"path/filepath"
+
+	"github.com/docker/docker/pkg/homedir"
+)
+
+var (
+	// ConfigDir configuration directory for falcoctl.
+	ConfigDir string
+	// FalcoctlPath path inside the configuration directory where the falcoctl stores its config files.
+	FalcoctlPath string
+	// IndexesFile name of the file where the indexes info is stored. It lives under FalcoctlPath.
+	IndexesFile string
+)
+
+func init() {
+	ConfigDir = filepath.Join(homedir.Get(), ".config")
+	FalcoctlPath = filepath.Join(ConfigDir, "falcoctl")
+	IndexesFile = filepath.Join(FalcoctlPath, "indexes.yaml")
+}
