+++ internal/consts/format
@@ -1,0 +1,6 @@
+package consts
+
+
+const(
+	TimeFormat   = "2006-01-02 15:04:05"
+)
